DipoleLattice.rt_P_dipole — Method

rt_P_dipole(pol::String, λ::Real, θ::Real, ϕ::Real, ɛμh::Vector{<:Real}, lattice::Lattice2D, αe::Array{<:Number, 2})

This function returns the field rflection and transmission coeficienets of a 2D array of dipoles with only electric dipole moment.

Arguments

pol::String: Polarization of incidence plane wave

λ::Real : The free space wavelength of the incidence light

θ::Real : The azimuthal angle of the incidence plane wave

ϕ::Real : The polar angle of the the incidence plane wave

ϵμh::Vector{<:Real} : Dielectric function and magnetic permitivities of the host medium. We use ϵμh = [ϵh, μh].

lattice::Lattice2D : Lattice objects that defines teh properties of 2D-lattice

αe : 3x3 electric polarizability tensor.

DipoleLattice.rt_M_dipole — Method

rt_M_dipole(pol::String, λ::Real, θ::Real, ϕ::Real, ɛμh::Vector{<:Real}, lattice::Lattice2D, αm::Array{<:Number, 2})

This function returns the field rflection and transmission coeficienets of a 2D array of dipoles with only magnetic dipole moments.

Arguments

pol::String: Polarization of incidence plane wave

λ::Real : The free space wavelength of the incidence light

θ::Real : The azimuthal angle of the incidence plane wave

ϕ::Real : The polar angle of the the incidence plane wave

ϵμh::Vector{<:Real} : Dielectric function and magnetic permitivities of the host medium. We use ϵμh = [ϵh, μh].

lattice::Lattice2D : Lattice objects that defines teh properties of 2D-lattice

αm : 3x3 magnetic polarizability tensor.

DipoleLattice.rt_PM_dipole — Method

rt_PM_dipole(pol::String, λ::Real, θ::Real, ϕ::Real, ɛμh::Vector{<:Real}, lattice::lattice2D, αem::Array{<:Number, 2})

This function returns the field rflection and transmission coeficienets of a 2D array of dipoles with both electric and magnetic dipole moments.

Arguments

pol::String: Polarization of incidence plane wave

λ::Real : The free space wavelength of the incidence light

θ::Real : The azimuthal angle of the incidence plane wave

ϕ::Real : The polar angle of the the incidence plane wave

ϵμh::Vector{<:Real} : Dielectric function and magnetic permitivities of the host medium. We use ϵμh = [ϵh, μh].

lattice::Lattice2D : Lattice objects that defines teh properties of 2D-lattice

αem : 6x6 magneto-electrical polarizability tensor.

DipoleLattice.DyadSum — Method

DyadSum(k::Number, rv::Vector{<:Real}, kp::Vector{<:Real}, L::Lattice2D, GorH::String; Nmax::Integer = 5)